# src/main/java/com/tenniscourts/guests/GuestDTO.java
package com.tenniscourts.guests;

import lombok.*;

import javax.validation.constraints.NotNull;

@AllArgsConstructor
@NoArgsConstructor
@Getter
@Setter
@Builder
public class GuestDTO {

    private Long id;

    @NotNull
    private String name;
}